+++ src/lib/components/v2/tasks/TaskGroupManageModal.svelte
@@ -19,7 +19,8 @@
 	let originalActive = $state(true);
 
 	let saving = $state(false);
-	let askConfirm = $state(false);
+	let askConfirmForDeactivate = $state(false);
+	let askConfirmForDelete = $state(false);
 
 	/** @type {import('$lib/components/common/StandardErrorAlert.svelte').default|undefined} */
 	let errorAlert = undefined;
@@ -30,7 +31,8 @@
 	export async function open(taskGroupToEdit) {
 		taskGroup = taskGroupToEdit;
 		originalActive = taskGroupToEdit.active;
-		askConfirm = false;
+		askConfirmForDeactivate = false;
+		askConfirmForDelete = false;
 		errorAlert?.hide();
 		modal?.show();
 	}
@@ -39,6 +41,7 @@
 	 * @param {boolean} active
 	 */
 	async function handleEditTaskGroup(active) {
+		errorAlert?.hide();
 		saving = true;
 		const headers = new Headers();
 		headers.append('Content-Type', 'application/json');
@@ -46,7 +49,7 @@
 		const response = await fetch(
 			`/api/${admin ? 'admin/' : ''}v2/task-group/${taskGroup?.id}/${
 				active ? 'reactivate' : 'deactivate'
-			}/`,
+			}`,
 			{
 				method: 'POST',
 				credentials: 'include',
@@ -54,14 +57,41 @@
 				body: JSON.stringify({})
 			}
 		);
-		saving = false;
 
+		await handleActivityResponse(response);
+	}
+
+	async function handleDeleteTaskGroup() {
+		errorAlert?.hide();
+		saving = true;
+		const headers = new Headers();
+		headers.append('Content-Type', 'application/json');
+
+		const response = await fetch(
+			`/api/${admin ? 'admin/' : ''}v2/task-group/${taskGroup?.id}/delete`,
+			{
+				method: 'POST',
+				credentials: 'include',
+				headers,
+				body: JSON.stringify({})
+			}
+		);
+
+		await handleActivityResponse(response);
+	}
+
+	/**
+	 * @param {Response} response
+	 */
+	async function handleActivityResponse(response) {
 		if (response.ok) {
 			const result = await response.json();
+			saving = false;
 			const page = admin ? `/v2/admin/task-groups/activities` : `/v2/tasks/activities`;
 			goto(`${page}?activity_id=${result.id}`);
 			modal?.hide();
 		} else {
+			saving = false;
 			errorAlert = displayStandardErrorAlert(
 				await getAlertErrorFromResponse(response),
 				'taskGroupManageError'
@@ -80,7 +110,7 @@
 		{/if}
 	{/snippet}
 	{#snippet body()}
-		{#if askConfirm}
+		{#if askConfirmForDeactivate}
 			<div class="row">
 				<div class="col">
 					<div class="alert alert-warning">
@@ -98,10 +128,31 @@
 						disabled={saving}
 					>
 						{#if saving}
-							<span class="spinner-border spinner-border-sm" role="status" aria-hidden="true"
-							></span>
+							<span class="spinner-border spinner-border-sm" role="status" aria-hidden="true">
+							</span>
 						{/if}
 						Confirm
+					</button>
+				</div>
+			</div>
+		{:else if askConfirmForDelete}
+			<div class="row">
+				<div class="col">
+					<div class="alert alert-warning">
+						<i class="bi bi-exclamation-triangle"></i>
+						Warning: this will both delete task-group metadata and (if applicable) the task-group folder.
+						Do you want to proceed?
+					</div>
+				</div>
+			</div>
+			<div class="row">
+				<div class="col">
+					<button class="btn btn-danger" onclick={() => handleDeleteTaskGroup()} disabled={saving}>
+						{#if saving}
+							<span class="spinner-border spinner-border-sm" role="status" aria-hidden="true">
+							</span>
+						{/if}
+						Confirm delete
 					</button>
 				</div>
 			</div>
@@ -114,7 +165,7 @@
 			<div class="row">
 				<div class="col">
 					{#if originalActive}
-						<button class="btn btn-primary" onclick={() => (askConfirm = true)}>
+						<button class="btn btn-primary" onclick={() => (askConfirmForDeactivate = true)}>
 							Deactivate task group
 						</button>
 					{:else}
@@ -124,8 +175,16 @@
 					{/if}
 				</div>
 			</div>
+			<div class="row">
+				<div class="col">
+					<hr />
+					<button class="btn btn-danger" onclick={() => (askConfirmForDelete = true)}>
+						Delete task group
+					</button>
+				</div>
+			</div>
 		{/if}
-		<span id="taskGroupManageError"></span>
+		<div id="taskGroupManageError" class="mt-3"></div>
 	{/snippet}
 	{#snippet footer()}
 		<button class="btn btn-secondary" data-bs-dismiss="modal">Cancel</button>