+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 # Unreleased
 
 * Added body request normalizer (\#825);
+* Fixed JSON schema validation during arguments import (\#827);
+* Updated slim-select version to support range selection (\#827);
+* Supported pre/post `pinned_package_versions` in task-collection (\#827);
 
 # 1.19.3